Imp: move EMS config to package
